requirements.txt
streamlit==1.28.0
tensorflow-cpu==2.10.0
Pillow==9.5.0
numpy==1.24.0
